major refactoring of css, for modularity purposes
