+++ src/components/About/About.module.scss
@@ -62,16 +62,6 @@
 
   .text {
     width: 70%;
-    // border: 1px solid yellowgreen;
-
-    p {
-      font-family: "Nunito", sans-serif;
-      font-size: 1.3rem;
-      font-weight: 400;
-      color: var(--gray-text);
-      text-align: justify;
-      margin: 20px 0 0 0;
-    }
 
     .highlight {
       color: var(--faded-golden);
