+++ src/components/About/DescriptionComponent/Description.module.scss
@@ -1,0 +1,14 @@
+p {
+  margin: 20px 0 0 0;
+  font-family: "Nunito", sans-serif;
+  font-size: 1.7rem;
+  font-weight: 400;
+  color: var(--gray-text);
+  text-align: justify;
+
+  .highlight {
+    color: var(--faded-golden);
+    // text-decoration: underline;
+    font-weight: 700;
+  }
+}
