+++ src/components/About/DescriptionComponent/DescriptionComponent.tsx
@@ -1,6 +1,7 @@
-import styles from "./About.module.scss";
+import styles from "./Description.module.scss";
 import { useEffect, useState } from "react";
-import fetchDataFromCMS from "../../utils/fetchDataFromCMS";
+import fetchDataFromCMS from "../../../utils/fetchDataFromCMS";
+import useAsyncError from "../../../hooks/useAsyncError";
 
 // helper type
 type TextChild = {
@@ -18,6 +19,7 @@ type Paragraph = {
 // fetches data + parses data + returns JSX
 const DescriptionComponent = () => {
   const [descriptionArr, setDescriptionArr] = useState<Paragraph[]>([]);
+  const throwError = useAsyncError();
 
   useEffect(() => {
     async function fetchData() {
@@ -26,10 +28,11 @@ const DescriptionComponent = () => {
         setDescriptionArr(response.data.attributes.description);
       } catch (error) {
         console.error(error);
+        throwError(error as Error);
       }
     }
     fetchData();
-  }, []);
+  }, [throwError]);
 
   return descriptionArr.map((block: Paragraph, index: number) => (
     <p key={index}>
