+++ src/components/Contact/Contact.module.scss
@@ -66,14 +66,6 @@
     width: 40%;
     gap: 2.5rem;
 
-    p {
-      font-family: "Nunito", sans-serif;
-      font-size: 1.7rem;
-      font-weight: 400;
-      color: var(--gray-text);
-      text-align: justify;
-    }
-
     .iconrow {
       display: flex;
       flex-direction: row;
