+++ src/components/Contact/TextComponent/Text.module.scss
@@ -1,0 +1,14 @@
+// p {
+//   font-family: "Nunito", sans-serif;
+//   font-size: 1.3rem;
+//   font-weight: 400;
+//   color: var(--gray-text);
+//   text-align: justify;
+//   margin: 0;
+// }
+
+// .highlight {
+//   color: var(--faded-golden);
+//   // text-decoration: underline;
+//   font-weight: 700;
+// }
