+++ src/components/Hero/Hero.module.scss
@@ -3,28 +3,28 @@
   flex-direction: column;
   margin-top: 5rem;
   // border: 1px solid white;
-}
 
-p {
-  font-family: "Fira Code", sans-serif;
-  font-size: 2rem;
-  font-weight: 400;
-  color: var(--faded-golden);
-  margin: 0;
-}
+  p {
+    font-family: "Fira Code", sans-serif;
+    font-size: 2rem;
+    font-weight: 400;
+    color: var(--faded-golden);
+    margin: 0;
+  }
 
-h1 {
-  font-family: "Poppins", sans-serif;
-  font-size: 5rem;
-  font-weight: 700;
-  color: var(--cream);
-  margin: 0;
-}
+  h1 {
+    font-family: "Poppins", sans-serif;
+    font-size: 5rem;
+    font-weight: 700;
+    color: var(--cream);
+    margin: 0;
+  }
 
-h2 {
-  font-family: "Poppins", sans-serif;
-  font-size: 2.5rem;
-  font-weight: 500;
-  color: var(--cream);
-  margin: 0;
+  h2 {
+    font-family: "Poppins", sans-serif;
+    font-size: 2.5rem;
+    font-weight: 500;
+    color: var(--cream);
+    margin: 0;
+  }
 }
